refactor: consolidate and separate interfaces
