+++ infra/staff/src/components/Disable2FA.tsx
@@ -10,14 +10,7 @@ import {
 import React, { useState } from "react";
 import { getEmail, getToken } from "../App"; // Import getEmail and getToken functions
 import { apiOrigin } from "../services/support";
-
-interface UserData {
-    subscription?: {
-        userID: string;
-        // Add other properties as per your API response structure
-    };
-    // Add other properties as per your API response structure
-}
+import type { UserData } from "../types";
 
 interface Disable2FAProps {
     open: boolean;
