+++ infra/staff/src/components/FamilyComponentTable.tsx
@@ -13,22 +13,9 @@ import * as React from "react";
 import { useEffect, useState } from "react";
 import { getEmail, getToken } from "../App";
 import { apiOrigin } from "../services/support";
+import type { FamilyMember, UserData } from "../types";
+import { formatUsageToGB } from "../utils/strings";
 import CloseFamily from "./CloseFamily";
-
-interface FamilyMember {
-    id: string;
-    email: string;
-    status: string;
-    usage: number;
-}
-
-interface UserData {
-    details: {
-        familyData: {
-            members: FamilyMember[];
-        };
-    };
-}
 
 const FamilyTableComponent: React.FC = () => {
     const [familyMembers, setFamilyMembers] = useState<FamilyMember[]>([]);
@@ -54,7 +41,7 @@ const FamilyTableComponent: React.FC = () => {
                 }
                 const userData = (await response.json()) as UserData; // Typecast to UserData interface
                 const members: FamilyMember[] =
-                    userData.details.familyData.members;
+                    userData.details?.familyData?.members || [];
                 setFamilyMembers(members);
             } catch (error) {
                 console.error("Error fetching family data:", error);
@@ -68,11 +55,6 @@ const FamilyTableComponent: React.FC = () => {
             console.error("Fetch data error:", error),
         );
     }, []);
-
-    const formatUsageToGB = (usage: number): string => {
-        const usageInGB = (usage / (1024 * 1024 * 1024)).toFixed(2);
-        return `${usageInGB} GB`;
-    };
 
     const handleOpenCloseFamily = () => {
         setCloseFamilyOpen(true);
