--- infra/staff/src/utils/strings.ts
@@ -1,13 +1,0 @@
-/**
- * User facing strings in the app.
- *
- * By keeping them separate, we make our lives easier if/when we need to
- * localize the corresponding pages. Right now, these are just the values in the
- * default language, English.
- */
-const S = {
-    hello: "Hello Ente!",
-    error_generic: "Oops, something went wrong.",
-};
-
-export default S;
